requirements.txt
asn1crypto==0.24.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cloudinary==1.10.0
cryptography==1.7.2
decorator==4.1.2
Django==1.11
django-encrypted-filefield==0.1.1
django-material==1.1.1
django-richtextfield==1.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
jwt==0.5.2
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
python-magic==0.4.12
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
validators==0.12.0
asn1crypto==0.24.0
astroid==1.6.1
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cloudinary==1.10.0
cryptography==1.7.2
decorator==4.1.2
Django==1.11
django-encrypted-filefield==0.1.1
django-material==1.1.1
django-richtextfield==1.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
isort==4.3.1
jwt==0.5.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
olefile==0.44
pbr==3.1.1
Pillow==4.3.0
psycopg2==2.7.3.2
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.5.3
pylint==1.8.2
pylint-django==0.9.0
pylint-plugin-utils==0.2.6
python-magic==0.4.12
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
validators==0.12.0
wrapt==1.10.11
